requirements.txt
google-genai
dotenv
spoonacular
flask